python_engine/clothing_recommender/requirements.txt
# Clothing Recommender - LLM/Agent dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# LLM/Agent Stack
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
langfuse>=2.0.0
openai>=1.0.0
tiktoken>=0.5.0

# Database
pymongo>=4.0.0
motor>=3.3.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.25.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
